광고 위젯 위치 조정 및 타이머 초기화 지연시간 변경, BreathingExerciseWidget 통합
