+++ lib/src/views/home/home_screen.dart
@@ -48,59 +48,7 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
   void initState() {
     super.initState();
     // 위젯이 처음 생성될 때 _showBottomSheet 함수를 호출합니다. (타이머가 실행 중이지 않을 때만)
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      // Read the timer state using ref
-      final isTimerRunning = ref.read(suDuckTimerProvider).isRunning;
-      if (!isTimerRunning) {
-        _showBottomSheet();
-      }
-    });
-  }
-
-  void _showBottomSheet() {
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled:
-          true, // Keep true if content might exceed screen height initially
-      builder: (BuildContext context) {
-        // Give the container a fixed height to prevent resizing
-        return Container(
-          height: 400.h, // Set a fixed height
-          width: double.infinity,
-          // color: Colors.transparent, // Color is often set by theme, can remove if default is fine
-          padding: EdgeInsets.symmetric(horizontal: 16.w, vertical: 16.h),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              // Replace the Text widget with the BreathingExerciseWidget
-              SizedBox(
-                  height: 160.h, // Add some space above the exercise
-                  child: const BreathingExerciseWidget()),
-              SizedBox(height: 16.h), // Add some space below the exercise
-              AdMobWidget.showBannerAd(80.h), // 실제 광고 위젯
-              SizedBox(height: 16.h),
-              SizedBox(
-                width: double.infinity,
-                child: ElevatedButton(
-                    onPressed: () {
-                      Navigator.pop(context);
-                    },
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: Colors.blue,
-                      shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(8),
-                      ),
-                      padding: EdgeInsets.symmetric(vertical: 12.h),
-                    ),
-                    child: Text(tr("Common.Close"),
-                        style:
-                            TextStyle(fontSize: 16.sp, color: Colors.white))),
-              ),
-            ],
-          ),
-        );
-      },
-    );
+    WidgetsBinding.instance.addPostFrameCallback((_) {});
   }
 
   @override
