+++ lib/src/views/home/home_seg1_screen.dart
@@ -17,7 +17,7 @@ class Seg1Screen extends StatelessWidget {
           SuDuckTimerWidget(),
           SuduckTimerControllBarWidget(),
           SubjectListWidget(),
-          AdMobWidget.showBannerAd(60),
+          AdMobWidget.showBannerAd(60, true),
         ],
       ),
     );
